Use a better example for git cherry-pick
It is more common to cherry-pick select changes from feature branches
into master than it is to do the other way around.

Raymond Chen of "The Old New Thing" disagrees that there is any good use
for cherry-pick and provides an alternate solution; his solution is
given in the notes.