Add illustrations to section pages
An illustration with the default repo (with a remote repository added)
was added to each section page. The parts that are not the focus for the
section are faded (e.g. for the Commits section, branches are faded).

gitdags was modified to include a "fade" style and new commands
`\gitbranchfade`, `\gitremotebranchfade` and `\gittagfade`.